simple max-k-sat

+++ MaxSATFormula.h
@@ -122,7 +122,7 @@ public:
   vec<double> occurance_list;
   vec<double> temp_occurance_list;
   vec<double> var_bias;
-  vector<int> m;
+  vector<uint64_t> m;
   vec<lbool> assignment;
   vector<uint32_t> weight_sampler;
   vector<uint32_t> weight_pool;
